requirements.txt
bidict==0.22.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4"
cffi==1.15.1 ; platform_python_implementation == "CPython" and sys_platform == "win32" and python_version >= "3.9" and python_version < "4.0"
charset-normalizer==2.1.1 ; python_version >= "3.9" and python_version < "4"
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
faker==15.3.4 ; python_version >= "3.9" and python_version < "4.0"
flask-socketio==5.3.2 ; python_version >= "3.9" and python_version < "4.0"
flask==2.2.2 ; python_version >= "3.9" and python_version < "4.0"
gevent==22.10.2 ; python_version >= "3.9" and python_version < "4.0"
greenlet==2.0.1 ; python_version >= "3.9" and python_version < "4.0" and platform_python_implementation == "CPython"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4"
importlib-metadata==5.2.0 ; python_version >= "3.9" and python_version < "3.10"
itsdangerous==2.1.2 ; python_version >= "3.9" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.1 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; platform_python_implementation == "CPython" and sys_platform == "win32" and python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-engineio==4.3.4 ; python_version >= "3.9" and python_version < "4.0"
python-socketio==5.7.2 ; python_version >= "3.9" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.9" and python_version < "4"
setuptools==65.6.3 ; python_version >= "3.9" and python_version < "4.0"
simple-websocket==0.9.0 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.13 ; python_version >= "3.9" and python_version < "4"
uwsgi==2.0.21 ; python_version >= "3.9" and python_version < "4.0"
werkzeug==2.2.2 ; python_version >= "3.9" and python_version < "4.0"
wsproto==1.2.0 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.11.0 ; python_version >= "3.9" and python_version < "3.10"
zope-event==4.6 ; python_version >= "3.9" and python_version < "4.0"
zope-interface==5.5.2 ; python_version >= "3.9" and python_version < "4.0"
